fix build warning

+++ include/paw3222.h
@@ -39,7 +39,7 @@ struct paw32xx_config {
     int16_t snipe_cpi;
     uint8_t snipe_divisor; // Additional precision divisor for snipe mode (default: 2)
     bool force_awake;
-    uint8_t rotation;      // Rotation angle (0, 90, 180, 270)
+    uint16_t rotation;     // Rotation angle (0, 90, 180, 270)
     uint8_t scroll_tick;   // Scroll tick threshold
 };
